Moved verif code types
