+++ lib/sendIdentityValidationCode.ts
@@ -2,21 +2,10 @@ import { FetchResult } from "apollo-link";
 import gql from "graphql-tag";
 
 import { fetchManagement } from "../src/utils/fetchManagement";
-import { IdentityInput } from "./@types/Identity";
-
-export type SendIdentityVerificationCodeInput = {
-  callbackUrl: string;
-  identity: IdentityInput;
-  localeCodeOverride?: string;
-  userId?: string;
-};
-
-export type SendIdentityValidationCodeResult = {
-  callbackUrl: string;
-  eventId: string;
-  localeCode: string;
-  nonce: string;
-};
+import {
+  SendIdentityValidationCodeResult,
+  SendIdentityVerificationCodeInput,
+} from "./@types/VerificationCode";
 
 const SEND_IDENTITY_VALIDATION_CODE = gql`
   mutation sendIdentityValidationCode(
